chore: MAT-153 미사용되는 mock 정리
