+++ src/mocks/byTeamType.ts
@@ -1,6 +1,18 @@
 import { TeamType } from '@/stores';
 
-import { dummyTeam1, dummyTeam2 } from './startingPlayers';
+export const dummyTeam1 = {
+  id: 1,
+  name: 'FC 울릉도',
+  teamColor: '#D91920',
+  logoImageUrl: '/images/team-logo.png',
+};
+
+export const dummyTeam2 = {
+  id: 2,
+  name: 'FC 독도',
+  teamColor: '#003A70',
+  logoImageUrl: '/images/team-logo.png',
+};
 
 export const mocked_getTeamByType = (teamType: TeamType) => {
   if (teamType === 'home') {
