+++ src/mocks/index.ts
@@ -1,4 +1,2 @@
-export * from './startingPlayers';
 export * from './api';
-export * from './startingPlayersOnGrid';
-export * from './getTeamByType';
+export * from './byTeamType';
