--- src/mocks/startingPlayers.ts
@@ -1,208 +1,0 @@
-export const dummyTeam1 = {
-  id: 1,
-  name: 'FC 울릉도',
-  teamColor: '#D91920',
-  logoImageUrl: '/images/team-logo.png',
-};
-
-export const dummyTeam2 = {
-  id: 2,
-  name: 'FC 독도',
-  teamColor: '#003A70',
-  logoImageUrl: '/images/team-logo.png',
-};
-
-export const dummyPlayerOfTeam1 = [
-  {
-    id: 1,
-    name: '김태인',
-    number: 1,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 2,
-    name: '김성빈',
-    number: 12,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 10,
-    fouls: 10,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 3,
-    name: '주유나',
-    number: 3,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 4,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 5,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 6,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 7,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 8,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 9,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 10,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 11,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 12,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 13,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 14,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 15,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-  {
-    id: 16,
-    name: '김주용',
-    number: 4,
-    position: 'FW',
-    profileImageUrl: '/images/player-profile.png',
-    goals: 10,
-    assists: 5,
-    fouls: 3,
-    yellowCards: 0,
-    redCards: 0,
-  },
-];
